Added uint8 cast to adjust_gamma function

+++ take_photo.py
@@ -76,10 +76,10 @@ def adjust_gamma(image, gamma=1.0):
     invGamma = 1.0 / gamma
     
     table = np.array([ ((i / 255.0) ** invGamma) * 255
-		for i in np.arange(0, 256)]).astype("uint8")
+		for i in np.arange(0, 256)])
     
     # apply gamma correction using the lookup table
-    return cv2.LUT(image, table)
+    return cv2.LUT(image.astype(np.uint8), table.astype(np.uint8))
 
 
 def crop(image):
